new: dev: usr: PETSc upgrade !version
 * PETSc-3.13.1 with HYPRE-2.18.2 and boost-1_73_0 is updated for CauldronV2
Related work items: #382794, #393082
================== COMMIT-END ========================

+++ development/applications/fastcauldron/src/Well.cpp
@@ -211,7 +211,7 @@ void Well::Save_Well_Data( Location& needle, const string& Related_Project_Name 
   
   int  Top_Z_Node_Index;
   int  Number_Of_Segment;
-  int  Number_Of_Formation_Layers = Basin_Model -> layers.size() - 2;
+  size_t  Number_Of_Formation_Layers = Basin_Model -> layers.size() - 2;
   int  Layer_It;
 
   bool IncludeGhosts = true;
